Add PICO-8 TAS and supporting code.
- PICO-8 runs at 30 FPS, so inputs are only updated every other frame.
- Ignore loading frames while single-stepping.
- Add a status readout for PICO-8 to Celeste.Studio including a generated level name.

The TAS is probably suboptimal in a lot of places. However, the red balloons spawn at random y-offsets, and there's no way to predict or manipulate the RNG without external tools, so the TAS attempts to hit the middle of each balloon even in cases where it could save a few frames by dashing sooner.